# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "More-Recipes provides a platform for users to share the awesome and exciting recipe ideas they have invented or learnt.Suppose a user comes up with a recipe, he/she can post it on More-Recipes and get feedback in form of reviews and votes from other users who explore that recipe. Users can also keep a list of their favorite recipes on application.",
    "version": "1.0.0",
    "title": "More Recipes"
  },
  "host": "more-recipes-ovie.herokuapp.com",
  "basePath": "/api/v1",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/user/category/{categoryId}": {
      "put": {
        "tags": [
          "Category"
        ],
        "summary": "Update category",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateCategory",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "categoryId",
            "in": "path",
            "description": "recipe that needs to be updated",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated category object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "category name changed SuccessFullly!",
            "schema": {
              "$ref": "#/definitions/UpdateCategory"
            }
          },
          "400": {
            "description": "Invalid category name input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Category Not Found!",
            "schema": {
              "$ref": "#/definitions/CategoryNotFound"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Category"
        ],
        "summary": "Delete category",
        "description": "This can only be done by the logged in user.",
        "operationId": "deleteCategory",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "categoryId",
            "in": "path",
            "description": "The category that needs to be deleted",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted SuccessFullly!",
            "schema": {
              "$ref": "#/definitions/DeleteCategory"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Category Not Found!",
            "schema": {
              "$ref": "#/definitions/CategoryNotFound"
            }
          }
        }
      }
    },
...
    "/favorite/{favoriteId}": {
      "delete": {
        "tags": [
          "Favorite"
        ],
        "summary": "UnFavorite recipe",
        "description": "This can only be done by the logged in user.",
        "operationId": "unFavoriteRecipe",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "favoriteId",
            "in": "path",
            "description": "The recipe that needs to be unfavorited",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Recipe Unfavorited SuccessFullly!",
            "schema": {
              "$ref": "#/definitions/unFavoriteRecipe"
            }
          },
          "401": {
            "description": "Invalid FavoriteId Input"
          },
          "404": {
            "description": "Recipe Not Found!",
            "schema": {
              "$ref": "#/definitions/RecipeNotFound"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Category": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
...
    "RecipeNotFound": {
      "type": "object",
      "required": [
        "status",
        "message"
      ],
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "message": {
          "type": "string",
          "example": "Recipe Not Found!"
        }
      }
    },
...
    "UpdateCategory": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string",
          "example": "category name changed SuccessFullly!"
        }
      }
    },
    "DeleteCategory": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string",
          "example": "Category deleted SuccessFullly!"
        }
      }
    },
...
    "CategoryNotFound": {
      "type": "object",
      "required": [
        "status",
        "message"
      ],
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "message": {
          "type": "string",
          "example": "Category Not Found!"
        }
      }
    },
...
    "unFavoriteRecipe": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string",
          "example": "Recipe Unfavorited SuccessFullly!"
        }
      }
    },
...
  },
...
}